Add collection of responses
The collect_responses method sends a message and waits for responses from any
number of subscribers. To handle these "collect calls" a special
on_collect can be used to register a handler. The method will then handle
much of the required protocol to make this work.

There is an option to supply a method to consider the collection complete depending on a message. (for example a 100 confidence of a query or something)